% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_cols.R
\name{my_cols}
\alias{my_cols}
\title{Benutzerdefinierte Farben als Hexcodes extrahieren}
\usage{
my_cols(...)
}
\arguments{
\item{...}{Farbbezeichnung(en) als Strings (character)}
}
\value{
Benannter Vektor mit Hexcodes unserer Farben
}
\description{
Falls keine Farben angegeben: gibt alle Farben zurück.
}
\examples{
my_cols()
my_cols("Earth yellow")
my_cols("Earth yellow", "Dartmouth green")
ggplot(mtcars, aes(x = hp, y = mpg)) +
geom_point(size = 2.5, col = my_cols("Dartmouth green"))
}
