% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_pal.R
\name{my_pal}
\alias{my_pal}
\title{Benutzerdefinierte Farbpaletten interpolieren}
\usage{
my_pal(palette = "alle", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Name der Farbpalette (String / character). Zur Zeit "alle", "gruen", "drei"}

\item{reverse}{Boolean: Palette umdrehen?}

\item{...}{Zusätzliche Argumente, die an grDevices::colorRampPalette() weitergegeben werden}
}
\description{
Benutzerdefinierte Farbpaletten interpolieren
}
\examples{
my_pal("alle")  # Liefert eine Funktion!
my_pal("alle")(8)
my_pal("alle")(8) |> colorspace::swatchplot()
my_pal("drei")(8) |> colorspace::swatchplot()
my_pal("gruen")(4) |> colorspace::swatchplot()
}
