% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_fill.R
\name{scale_color_meinefa}
\alias{scale_color_meinefa}
\title{Color scale-Konstruktor für benutzerdefinierte Farben}
\usage{
scale_color_meinefa(palette = "alle", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Name (String) der Farbpalette}

\item{discrete}{Boolean: Farbskala diskret oder nicht?}

\item{reverse}{Boolean: Skala umdrehen?}

\item{...}{Weitere Parameter, die an discrete_scale() oder scale_color_gradientn() weitergegeben werden, je nach discrete TRUE / FALSE}
}
\description{
Color scale-Konstruktor für benutzerdefinierte Farben
}
\examples{
ggplot(mtcars, aes(x = hp, y = mpg, color = factor(cyl))) +
  geom_point(size = 3) +
  scale_color_meinefa()

ggplot(mtcars, aes(x = hp, y = mpg, color = factor(cyl))) +
  geom_point(size = 3) +
  scale_color_meinefa(palette = "drei")
}
