% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_fill.R
\name{scale_fill_meinefa}
\alias{scale_fill_meinefa}
\title{Fill scale-Konstruktor f체r benutzerdefinierte Farben}
\usage{
scale_fill_meinefa(palette = "alle", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Name (String) der Farbpalette}

\item{discrete}{Boolean: Farbskala diskret oder nicht?}

\item{reverse}{Boolean: Skala umdrehen?}

\item{...}{Weitere Parameter, die an discrete_scale() oder scale_color_gradientn() weitergegeben werden, je nach discrete TRUE / FALSE}
}
\description{
Fill scale-Konstruktor f체r benutzerdefinierte Farben
}
\examples{
ggplot(mtcars, aes(x = factor(cyl), fill = interaction(am, gear, sep = "/"))) +
  geom_bar(position = position_stack()) +
  scale_fill_meinefa(name = "Automatik / G채nge")

ggplot(mtcars, aes(x = factor(cyl), fill = interaction(am, gear, sep = "/"))) +
  geom_bar(position = position_stack()) +
  scale_fill_meinefa(name = "Automatik / G채nge", palette = "drei")  # Ohne grau
}
